ryanrealestate/requirements.txt
Django==1.9
django-appconf==1.0.1
django-classy-tags==0.6.2
django-easy-maps==0.9.2
django-mptt==0.7.4
django-mptt-urls==2.0.3
geopy==1.11.0
Pillow==3.0.0
psycopg2==2.6.1
six==1.10.0
